backend/requirements.txt
# 核心框架依赖
fastapi==0.103.0
uvicorn==0.23.2

# 数据库连接
mysql-connector-python==8.2.0

# 数据验证
pydantic==2.4.2
pydantic-settings==2.0.3

# 生物信息学处理
biopython==1.85

# 生产环境依赖
gunicorn==20.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
python-magic==0.4.27